backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-settings==2.6.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
openai==1.57.2
stripe==11.2.0
httpx==0.28.1
python-multipart==0.0.18
email-validator==2.3.0
PyMuPDF==1.26.7
Pillow==12.0.0
openpyxl==3.1.5
# PDF Processing
PyMuPDF==1.25.7
Pillow==11.1.0
# Excel/Data processing
openpyxl==3.1.5
pandas==2.2.3
